# src/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Notification API",
        "description": "Multi-channel (SMS / Email / Push) notification dispatch service with batch support, idempotency and async processing.",
        "contact": {
            "name": "In01",
            "email": "support@in01.local"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8080",
            "description": "Local"
        }
    ],
    "paths": {
...
        "/api/notifications/sms": {
            "post": {
                "tags": [
                    "Notifications"
                ],
                "summary": "Send an SMS notification",
                "operationId": "baef60064fce8b4571578a97c465ca3f",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/SendSmsRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "202": {
                        "description": "Accepted for async dispatch",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/AcceptedNotification"
                                }
                            }
                        }
                    },
                    "409": {
                        "description": "Duplicate (idempotency hit)",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/DuplicateNotificationError"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/notifications/email": {
            "post": {
                "tags": [
                    "Notifications"
                ],
                "summary": "Send an email notification",
                "operationId": "e7e69e81fa0dffbe9414a8e31ad51df1",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/SendEmailRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "202": {
                        "description": "Accepted for async dispatch",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/AcceptedNotification"
                                }
                            }
                        }
                    },
                    "409": {
                        "description": "Duplicate (idempotency hit)",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/DuplicateNotificationError"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/notifications/push": {
            "post": {
                "tags": [
                    "Notifications"
                ],
                "summary": "Send a push notification",
                "operationId": "0704bfd3a5863c5f4d8b20f28a230735",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/SendPushRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "202": {
                        "description": "Accepted for async dispatch",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/AcceptedNotification"
                                }
                            }
                        }
                    },
                    "409": {
                        "description": "Duplicate (idempotency hit)",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/DuplicateNotificationError"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/notifications/bulk": {
            "post": {
                "tags": [
                    "Notifications"
                ],
                "summary": "Send a batch of mixed-channel notifications (max 1000)",
                "operationId": "c00b97defc7548ab5dca396d44218f36",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/BulkRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "202": {
                        "description": "Batch accepted",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/BulkResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/notifications": {
            "get": {
                "tags": [
                    "Notifications"
                ],
                "summary": "List notifications (paginated, filterable)",
                "operationId": "6d56a555955089a73fdf3f7229e8ead8",
                "parameters": [
                    {
                        "name": "status",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "$ref": "#/components/schemas/NotificationStatus"
                        }
                    },
                    {
                        "name": "channel",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "$ref": "#/components/schemas/NotificationChannel"
                        }
                    },
                    {
                        "name": "batch_id",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    },
                    {
                        "name": "from",
                        "in": "query",
                        "description": "created_at lower bound (YYYY-MM-DD)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "to",
                        "in": "query",
                        "description": "created_at upper bound (YYYY-MM-DD)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 20,
                            "maximum": 100,
                            "minimum": 1
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 1,
                            "minimum": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/NotificationCollection"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/notifications/{id}": {
            "get": {
                "tags": [
                    "Notifications"
                ],
                "summary": "Get a single notification with channel-specific detail",
                "operationId": "4b6792a65cdd7f79b06bff8158d28094",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/NotificationResource"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/notifications/cancel/{id}": {
            "post": {
                "tags": [
                    "Notifications"
                ],
                "summary": "Cancel a pending or processing notification",
                "operationId": "ecaa476d64320eca50185502b2933dae",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Cancelled",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "string",
                                            "format": "uuid"
                                        },
                                        "status": {
                                            "$ref": "#/components/schemas/NotificationStatus"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "409": {
                        "description": "Not cancellable (terminal status)",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "required": [
                                        "message"
                                    ],
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Notification with status 'sent' cannot be cancelled."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/notifications/cancel/batch/{batchId}": {
            "post": {
                "tags": [
                    "Notifications"
                ],
                "summary": "Cancel all cancellable notifications in a batch",
                "operationId": "9bcda4c7137d5c2709d580fc0dbcf518",
                "parameters": [
                    {
                        "name": "batchId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "IDs of cancelled notifications",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "cancelled": {
                                            "type": "array",
                                            "items": {
                                                "type": "string",
                                                "format": "uuid"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "ValidationError": {
                "required": [
                    "message",
                    "errors"
                ],
                "properties": {
                    "message": {
                        "type": "string",
                        "example": "The given data was invalid."
                    },
                    "errors": {
                        "type": "object",
                        "example": {
                            "recipient": [
                                "The recipient field is required."
                            ]
                        },
                        "additionalProperties": {
                            "type": "array",
                            "items": {
                                "type": "string"
                            }
                        }
                    }
                },
                "type": "object"
            },
            "ErrorResponse": {
                "required": [
                    "message"
                ],
                "properties": {
                    "message": {
                        "type": "string",
                        "example": "Resource not found."
                    }
                },
                "type": "object"
            },
            "DuplicateNotificationError": {
                "required": [
                    "message",
                    "existing_notification_id"
                ],
                "properties": {
                    "message": {
                        "type": "string",
                        "example": "Notification already exists with this idempotency key"
                    },
                    "existing_notification_id": {
                        "type": "string",
                        "format": "uuid"
                    }
                },
                "type": "object"
            },
            "AcceptedNotification": {
                "required": [
                    "id",
                    "status",
                    "created_at"
                ],
                "properties": {
                    "id": {
                        "type": "string",
                        "format": "uuid",
                        "example": "0a3b2f9c-2bb0-4d7c-8f0f-7c1d5e2c1f0a"
                    },
                    "status": {
                        "$ref": "#/components/schemas/NotificationStatus"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
            "NotificationChannel": {
                "type": "string",
                "enum": [
                    "sms",
                    "email",
                    "push"
                ]
            },
            "NotificationStatus": {
                "type": "string",
                "enum": [
                    "pending",
                    "processing",
                    "sent",
                    "failed",
                    "cancelled"
                ]
            },
            "SendEmailRequest": {
                "required": [
                    "recipient",
                    "subject",
                    "body"
                ],
                "properties": {
                    "recipient": {
                        "type": "string",
                        "format": "email",
                        "example": "mustafa@example.com"
                    },
                    "subject": {
                        "type": "string",
                        "example": "Siparişiniz Onaylandı",
                        "maxLength": 255
                    },
                    "body": {
                        "type": "string",
                        "example": "Sayın müşterimiz, siparişiniz başarıyla onaylanmıştır.",
                        "maxLength": 10000
                    },
                    "priority": {
                        "description": "Message Priority",
                        "type": "string",
                        "enum": [
                            "Low",
                            "Medium",
                            "High"
                        ]
                    },
                    "scheduled_at": {
                        "description": "Scheduled delivery time",
                        "type": "string",
                        "example": "2026-07-01 09:00",
                        "nullable": true
                    }
                },
                "type": "object"
            },
...
            "SendSmsRequest": {
                "required": [
                    "recipient",
                    "content"
                ],
                "properties": {
                    "recipient": {
                        "description": "E.164 formatted phone number",
                        "type": "string",
                        "example": "+905551234567",
                        "maxLength": 16
                    },
                    "content": {
                        "type": "string",
                        "example": "Merhaba, siparişiniz hazır.",
                        "maxLength": 160
                    },
                    "priority": {
                        "description": "Message Priority",
                        "type": "string",
                        "enum": [
                            "Low",
                            "Medium",
                            "High"
                        ]
                    },
                    "scheduled_at": {
                        "description": "Scheduled delivery time",
                        "type": "string",
                        "example": "2026-07-01 09:00",
                        "nullable": true
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}